app/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
ecdsa==0.18.0
fastapi==0.101.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy==1.25.2
outcome==1.2.0
packaging==23.1
pandas==2.1.0
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
psycopg2==2.9.7
pyasn1==0.5.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
PyJWT==2.8.0
pytest==7.4.1
pytest-asyncio==0.21.1
pytest-trio==0.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
starlette==0.27.0
trio==0.22.2
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.2
watchfiles==0.19.0
websockets==11.0.3
